feat: initial changes for image and containers
