--- src/features/domain-preview/DomainPreview.module.scss
@@ -1,84 +1,0 @@
-@import '../../../node_modules/@zero-tech/zui/styles/_borders.scss';
-@import '../../../node_modules/@zero-tech/zui/styles/_colors.scss';
-@import '../../../node_modules/@zero-tech/zui/styles/_typography.scss';
-
-.Container {
-	display: flex;
-	flex-direction: column;
-	padding: 1rem 0;
-
-	.Banner {
-		width: 100%;
-		height: 15rem;
-		border-radius: calc($radius-primary / 2);
-
-		.Media {
-			height: 100%;
-			width: 100%;
-			border-radius: inherit;
-		}
-
-		&.isNFTView {
-			height: 23.125rem;
-			display: flex;
-			justify-content: center;
-		}
-	}
-
-	.Content {
-		margin-top: -3.75rem;
-		padding-bottom: 0.5rem;
-
-		&.isNFTView {
-			margin-top: 0;
-		}
-
-		.Icon {
-			height: 7.5rem;
-			width: 7.5rem;
-			margin-left: 0.5rem;
-			border-radius: 50%;
-			box-shadow: 0 0 1rem rgba($black, 0.5);
-			z-index: 2;
-		}
-
-		.TextContainer {
-			margin-top: 1.5rem;
-
-			.Title {
-				font-size: $size-large;
-				margin: 0;
-			}
-
-			.Description {
-				margin: 0.5rem 0 0 0;
-			}
-
-			.FlexRow {
-				display: flex;
-				justify-content: space-between;
-				align-items: center;
-
-				.MemberContainer {
-					display: flex;
-					padding: 0;
-					margin: 1.5rem 0 0 0;
-					list-style: none;
-
-					> *:not(:first-child) {
-						margin-left: 2rem;
-					}
-				}
-			}
-
-			.LinkContainer {
-				margin-top: 1rem;
-
-				.Link {
-					color: $blue;
-					font-size: $size-tiny;
-				}
-			}
-		}
-	}
-}
