--- src/features/domain-preview/index.ts
@@ -1,1 +1,0 @@
-export * from './DomainPreview';
