+++ src/features/view-nft/DetailsCard/index.ts
@@ -1,0 +1,1 @@
+export * from './DetailsCard';
