+++ src/features/view-nft/DetailsCard/ui/CTAContainer/index.ts
@@ -1,0 +1,1 @@
+export * from './CTAContainer';
