+++ src/features/view-nft/DetailsCard/ui/Options/Options.module.scss
@@ -1,5 +1,5 @@
-@import '../../../../node_modules/@zero-tech/zui/styles/_borders.scss';
-@import '../../../../node_modules/@zero-tech/zui/styles/_colors.scss';
+@import '../../../../../../node_modules/@zero-tech/zui/styles/_borders.scss';
+@import '../../../../../../node_modules/@zero-tech/zui/styles/_colors.scss';
 
 .Container {
 	display: flex;
