+++ src/features/view-nft/DetailsCard/ui/Options/index.ts
