+++ src/features/view-nft/DetailsCard/ui/index.ts
@@ -1,0 +1,2 @@
+export * from './CTAContainer';
+export * from './Options';
