+++ src/features/view-nft/index.ts
@@ -1,5 +1,7 @@
 export * from './Action';
 export * from './Actions';
+export * from './Banner';
+export * from './DetailsCard';
 export * from './HistoryItem';
 export * from './HistoryList';
 export * from './NFTMetrics';
