+++ src/features/view-subdomains/Banner/Banner.module.scss
@@ -1,0 +1,34 @@
+@import '../../../../node_modules/@zero-tech/zui/styles/_borders.scss';
+@import '../../../../node_modules/@zero-tech/zui/styles/_colors.scss';
+
+.Container {
+	display: flex;
+	flex-direction: column;
+	margin: 6rem 0 4rem;
+
+	.Banner {
+		width: 100%;
+		height: 15rem;
+		border-radius: calc($radius-primary / 2);
+
+		.Media {
+			height: 100%;
+			width: 100%;
+			border-radius: inherit;
+		}
+
+		.Thumbnail {
+			margin-top: -3.75rem;
+			padding-bottom: 2.5rem;
+
+			.Icon {
+				height: 7.5rem;
+				width: 7.5rem;
+				margin-left: 0.5rem;
+				border-radius: 50%;
+				box-shadow: 0 0 1rem rgba($black, 0.5);
+				z-index: 2;
+			}
+		}
+	}
+}
