+++ src/features/view-subdomains/Banner/index.ts
@@ -1,0 +1,1 @@
+export * from './Banner';
