+++ src/features/view-subdomains/DetailsCard/DetailsCard.module.scss
@@ -1,0 +1,26 @@
+@import '../../../../node_modules/@zero-tech/zui/styles/_colors.scss';
+@import '../../../../node_modules/@zero-tech/zui/styles/_typography.scss';
+
+.Container {
+	padding: 1.5rem 0;
+
+	.Content {
+		.Title {
+			font-size: $size-large;
+			margin: 0;
+		}
+
+		.Description {
+			margin: 0.5rem 0 0 0;
+		}
+
+		.LinkContainer {
+			margin-top: 1rem;
+
+			.Link {
+				color: $blue;
+				font-size: $size-tiny;
+			}
+		}
+	}
+}
