+++ src/features/view-subdomains/index.ts
@@ -1,3 +1,5 @@
+export * from './Banner';
+export * from './DetailsCard';
 export * from './SubdomainMetrics';
 export * from './SubdomainTable';
 export * from './SubdomainTableRow';
